Спринт 9: Убрать main из FileBackedTaskManager
